server: move source types out of facade
