+++ crates/server/src/source.rs
@@ -24,8 +24,6 @@
 //! Reactivity stays out (this crate is Dioxus-free): callers bump generations /
 //! nudge the sync task after a successful op.
 
-use std::fmt;
-
 use async_trait::async_trait;
 
 use config::{AppConfig, MusicService, Source};
@@ -35,230 +33,10 @@ use crate::jellyfin::JellyfinClient;
 use crate::server_ops::ServerConn;
 use crate::subsonic::SubsonicClient;
 use crate::ytmusic::YouTubeMusicClient;
-use crate::ytmusic::player::AudioFormat;
 
 pub mod capabilities;
-
-/// Why a [`MediaSource`] operation failed, classified so the UI can react
-/// differently (toast vs re-auth prompt vs "not supported") instead of pattern-
-/// matching an opaque string.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SourceError {
-    /// This source doesn't offer the operation (capability-gated). Carries the
-    /// op name; unreachable when the UI honours [`Capabilities`].
-    Unsupported(&'static str),
-    /// No usable connection to the remote (offline / creds-less stand-in).
-    Connectivity,
-    /// The remote rejected the credentials — re-auth needed.
-    Auth,
-    /// The caller passed something the operation can't act on (e.g. a track
-    /// missing the id a remote needs).
-    InvalidInput(String),
-    /// An underlying client/DB error, kept as its message.
-    Backend(String),
-}
-
-impl SourceError {
-    /// Construct an [`Unsupported`](SourceError::Unsupported) for `op`.
-    pub fn unsupported(op: &'static str) -> Self {
-        SourceError::Unsupported(op)
-    }
-}
-
-impl fmt::Display for SourceError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            SourceError::Unsupported(op) => write!(f, "this source doesn't support {op}"),
-            SourceError::Connectivity => f.write_str("the server has no active connection"),
-            SourceError::Auth => {
-                f.write_str("this source isn't signed in — open Settings to re-sign in")
-            }
-            SourceError::InvalidInput(m) | SourceError::Backend(m) => f.write_str(m),
-        }
-    }
-}
-
-impl std::error::Error for SourceError {}
-
-impl From<String> for SourceError {
-    fn from(m: String) -> Self {
-        SourceError::Backend(m)
-    }
-}
-
-impl From<db::DbError> for SourceError {
-    fn from(e: db::DbError) -> Self {
-        SourceError::Backend(e.to_string())
-    }
-}
-
-/// What a source can do with playlists — a genuine three-way split (local and
-/// most servers can reorder; YT Music can add/remove but not reorder; an
-/// offline/creds-less server can do neither). `Ord` so the UI can gate with
-/// `caps.playlists >= PlaylistOps::AddRemove`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum PlaylistOps {
-    /// No playlist mutation (no live connection).
-    None,
-    /// Create / add / remove, but no reordering.
-    AddRemove,
-    /// Full editing, including reordering entries.
-    Reorder,
-}
-
-/// How a source's artist view behaves — the artists tab routes on this instead
-/// of hardcoding a service. (`Unsupported` isn't modelled yet: every source has
-/// some artist view, and an unconstructed variant would be dead code.)
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ArtistView {
-    /// A library-driven page: a grid of artists → their tracks. Local, Jellyfin,
-    /// Subsonic — and a creds-less/offline server (cached artists).
-    Library,
-    /// A rich remote profile (banner, top songs, albums, related) when an artist
-    /// is selected; the library grid otherwise. YT Music.
-    Remote,
-}
-
-/// How favorites are fetched. `Instant` returns the whole set in one call
-/// ([`fetch_favorites`]); `Paginated` walks pages ([`fetch_favorites_page`]) so
-/// the UI can stream rows + show progress (YT's 800-song liked library).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FavoritesSync {
-    Instant,
-    Paginated,
-}
-
-/// How the album page renders, routed on this instead of hardcoding a service.
-/// `Standard` is the library-driven [`TrackListView`]; `YtMusic` is the rich
-/// catalog-remote page (header + on-demand full track list resolved from the
-/// remote). Catalog remotes that store albums by a title+artist hash with no
-/// full local track list select `YtMusic`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AlbumType {
-    /// Standard album page from the local/library tracks.
-    Standard,
-    /// YT-Music-style page: instant header, full track list pulled on demand.
-    YtMusic,
-}
-
-/// One page of paginated favorites: the tracks plus the cursor for the next page
-/// (`None` once exhausted). Cross-page dedup is the caller's job.
-pub struct FavoritesPage {
-    pub tracks: Vec<reader::Track>,
-    pub next: Option<String>,
-}
-
-/// One page of a playlist's entries: the tracks plus the cursor for the next page
-/// (`None` once exhausted). Cross-page dedup is the caller's job. Sources that
-/// can't page return everything in a single page (the default impl does this).
-pub struct PlaylistPage {
-    pub tracks: Vec<reader::Track>,
-    pub next: Option<String>,
-}
-
-/// A full remote library pull: albums, tracks, and artist `(name, image_url)`
-/// pairs, already transformed into the generic model types (each source applies
-/// its own id-prefix / cover encoding). The caller persists + prunes; keeping the
-/// write side out of the source preserves the chunked upsert/bump streaming.
-#[derive(Default)]
-pub struct LibrarySnapshot {
-    pub albums: Vec<reader::Album>,
-    pub tracks: Vec<reader::Track>,
-    pub artist_images: Vec<(String, String)>,
-}
-
-/// What a source supports, declared by each impl and read off the resolved
-/// source. Source-agnostic pages gate their few divergent affordances on these
-/// flags rather than on the source kind or service.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Capabilities {
-    /// Edit a track's tags (writes to the file) — local only.
-    pub edit_tags: bool,
-    /// Delete a track from disk — local only.
-    pub delete_from_disk: bool,
-    /// Scan/import filesystem folders — local only.
-    pub scan_folders: bool,
-    /// Organise playlists into folders (local) vs a flat remote list (servers).
-    pub folders: bool,
-    /// Manual refresh / background reconcile against a remote — servers.
-    pub sync: bool,
-    /// Download tracks for offline — servers.
-    pub downloads: bool,
-    /// A discover/recommendations surface — YT (and future remotes).
-    pub discover: bool,
-    /// Start a radio/mix from a seed track — YT.
-    pub radio: bool,
-    /// Playlist mutation level.
-    pub playlists: PlaylistOps,
-    /// How the artists tab renders a selected artist.
-    pub artist_view: ArtistView,
-    /// How the album page renders.
-    pub albums: AlbumType,
-    /// How favorites sync — one shot vs paginated (streamed with progress).
-    pub favorites_sync: FavoritesSync,
-}
-
-/// What credential validation concluded. `Unreachable` ≠ `Expired`: a network
-/// blip must not reprompt sign-in — only a real auth rejection does.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AuthOutcome {
-    Valid,
-    Expired,
-    Unreachable,
-}
-
-/// A resolved playable stream for one item. `format`/`user_agent`/`duration_secs`
-/// /`bitrate` are populated only where the source provides them (YT's deciphered
-/// stream carries all four; a plain progressive URL leaves them `None`).
-pub struct StreamInfo {
-    pub url: String,
-    pub format: Option<(AudioFormat, bool)>,
-    pub user_agent: Option<String>,
-    pub duration_secs: Option<u64>,
-    pub bitrate: Option<u32>,
-    /// Total byte length of the stream, when the source reports it (YT does — the
-    /// offline downloader uses it for progress). `None` for plain URL sources.
-    pub content_length: Option<u64>,
-}
-
-/// A remote playlist's listing metadata (no entries). `image_tag` is the remote
-/// cover reference (Jellyfin Primary tag / encoded YT thumbnail URL), `None` for
-/// sources without one.
-pub struct PlaylistMeta {
-    pub id: String,
-    pub name: String,
-    pub image_tag: Option<String>,
-}
-
-/// A full remote album — header metadata + every track — for the YT-Music-style
-/// album page, transformed into the generic model types. Source-agnostic: the
-/// album page renders this without reaching into any per-service catalog struct.
-/// `browse_id`/`audio_playlist_id` are opaque remote ids (the catalog uses them
-/// to re-fetch / start the album radio); `None` where the source has none.
-#[derive(Debug, Clone, PartialEq)]
-pub struct RemoteAlbum {
-    pub browse_id: String,
-    pub title: String,
-    pub artist: Option<String>,
-    pub year: Option<String>,
-    pub thumbnail: Option<String>,
-    pub audio_playlist_id: Option<String>,
-    pub tracks: Vec<reader::Track>,
-}
-
-impl From<crate::ytmusic::discover::YtAlbum> for RemoteAlbum {
-    fn from(a: crate::ytmusic::discover::YtAlbum) -> Self {
-        Self {
-            browse_id: a.browse_id,
-            title: a.title,
-            artist: a.artist,
-            year: a.year,
-            thumbnail: a.thumbnail,
-            audio_playlist_id: a.audio_playlist_id,
-            tracks: a.tracks,
-        }
-    }
-}
+mod types;
+pub use types::*;
 
 /// The source-agnostic backend the app drives. Impls supply
 /// [`source`](Self::source) + [`db`](Self::db) (the seam the default methods
